deeptsf_backend/celery_DeepTSF/requirements.txt
celery==5.4.0
fastapi==0.110.1
pretty-errors==1.2.25
u8darts==0.24.0
pandas==2.2.1
matplotlib==3.8.3
mlflow==2.11.3
click==8.1.7
tqdm==4.66.2
holidays==0.46
pytz==2024.1
numpy==1.26.4
urllib3==2.2.1
python-dotenv==1.0.1
minio==7.2.5
